OLMIS-2604: create framework for csv upload and use it for CCE cataog
One can create a CCE catalog in the system of all the possible CCE in system so that each CCE technician can add/register their specific equipment inventory.
Deletion of CCE is not allowed; the CSV only determines what to add or update. If a CCE is in the system but is not in the CSV, it is unchanged.
Added an endpoint to pass in a CSV for import into the database